propagator: always emit the ready signal even if it was already emitted
In case of parallel chunking, new jobs are not started when it is emited for
the parallel chunks are already taking all the lines.  It was emitted again when
the chunks were not going in parallel anymore but this has no effect.

I think it's safe to always emit it, because even if we start several neon job
in parellel because of that, they should queue in their thread

This change do not need to go to mirall master because the scheduling is different